Deepak -- STABLE -- imageSeriesVisualization
Changed genImageMaskOverlay.m to take multiple masks as input

fixed a few matlab 2012b specific bugs in imseriesmaskshow.m and imseriesmaskshowrgb.m. Shouldnt effect its function on other versions

git-svn-id: https://orchestra.med.harvard.edu/svn/lccb/matlab/projects/common/main@9397 0bcb941b-74d1-0310-bb26-e6636707f451

+++ graphics/imageSeriesVisualization/imseriesmaskshowrgb.m
@@ -624,7 +624,7 @@ end
 %%
 function [ intensityRange ] = ComputeImageDynamicRange( im, cover_percent )
 
-    [p,x] = hist( im, 255 );   
+    [p,x] = hist( double(im), 255 );   
     p = p / sum(p);
     
     min_xlow = [];
